# techsol-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TechnologySolutions Onboarding API",
        "description": "Documentação da API para o sistema de onboarding da TechnologySolutions",
        "contact": {
            "email": "kaeffea@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autentica um usuário e retorna um token de acesso",
                "description": "Autentica um usuário com CPF e senha.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Credenciais do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "password"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "format": "cpf",
                                        "example": "111.222.333-44"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login bem-sucedido!"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|Abcde..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou credenciais incorretas"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registra um novo usuário a partir de um convite",
                "description": "Finaliza o cadastro de um novo colaborador usando um token de convite válido.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Dados do novo colaborador",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "name",
                                    "cpf",
                                    "password"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "token_recebido_no_email"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Nome Completo"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "NovaSenhaForte123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário cadastrado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token de convite inválido ou expirado"
                    },
                    "422": {
                        "description": "Erro de validação dos dados"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        }
    ]
}